changed var name

+++ convnet_cifar.py
@@ -2,6 +2,7 @@ import cifar10_input
 cifar10_input.maybe_download_and_extract()
 
 import tensorflow as tf
+import numpy as np
 import time, os
 
 # Architecture
@@ -170,7 +171,7 @@ if __name__ == '__main__':
                     # Loop over all batches
                     for i in range(total_batch):
                         # Fit training using batch data
-                        _, new_cost = sess.run([train_op, cost], feed_dict={train_or_eval: 1, keep_prob: 0.5})
+                        _, new_cost = sess.run([train_op, cost], feed_dict={train_or_eval: np.int32(1), keep_prob: 0.5})
                         # Compute average loss
                         avg_cost += new_cost/total_batch
                         print "Epoch %d, minibatch %d of %d. Average cost = %0.4f." %(epoch, i, total_batch, avg_cost)
